requirements.txt
PyYAML>=6.0
Jinja2>=3.1.0
requests>=2.31.0
playwright>=1.40.0
